+++ backend/bait-api/app/Modules/Multimedia/Http/Controllers/PostReactionController.php
@@ -91,7 +91,7 @@ class PostReactionController extends Controller
                                   ->where('post_id', $post_id)
                                   ->first();
             $reaction->update(['reaction_type_id' => $reaction_type_id]);
-            event(new NewReactionEvent($reaction));
+            //event(new NewReactionEvent($reaction));
             return response()->json(new PostReactionResource($reaction->load('reactionType')), 200);
         }